+++ src/square/number.rs
